feat: also hide highlighting for first guess
