+++ Prolangle.Abstractions/Snippets/IColorfulTypeHighlighter.cs
@@ -1,0 +1,6 @@
+using Blism;
+using Prolangle.Abstractions.Services;
+
+namespace Prolangle.Abstractions.Snippets;
+
+public interface IColorfulTypeHighlighter : ITokenTypeHighlighter<GeneralTokenType>;
