+++ Prolangle.Abstractions/Snippets/IColorlessTypeHighlighter.cs
@@ -1,0 +1,6 @@
+using Blism;
+using Prolangle.Abstractions.Services;
+
+namespace Prolangle.Abstractions.Snippets;
+
+public interface IColorlessTypeHighlighter : ITokenTypeHighlighter<GeneralTokenType>;
