+++ Prolangle/Pages/Snippets.razor
@@ -77,6 +77,7 @@
 }
 
 <ConcealedSourceCode Snippet="@Controller.Snippet" RevealedPercent="@Controller.RevealedPercent"
+                     RevealHighlighting="@Controller.RevealHighlighting"
                      Style="padding: 0.5rem; border-radius: 0.5rem;"/>
 
 <div style="display: flex; flex-wrap: wrap; flex-direction: row; gap: 1em; margin-top: 1rem;">
